chore(changelog): Update changelog for version 3.6.1
Fix issue with `disc` calculation in the two_nodes_gagge model and
limit its value to 6. Improve naming of sweat-related outputs in the
`gagge` model and enhance documentation for the `disc` function.
Update PMV ASHRAE model to return the `compliance` boolean value
according to ASHRAE 55:2023 standard. Improve formatting of model
outputs to the console.